+++ include/picongpu/plugins/particleCalorimeter/ParticleCalorimeterFunctors.hpp
@@ -111,13 +111,13 @@ namespace picongpu
                 // yaw
                 int32_t yawBin = calorimeterPos.x() * static_cast<float_X>(numBinsYaw);
                 // catch out-of-range values
-                yawBin = yawBin >= numBinsYaw ? numBinsYaw - 1 : yawBin;
+                yawBin = yawBin >= static_cast<int32_t>(numBinsYaw) ? numBinsYaw - 1 : yawBin;
                 yawBin = yawBin < 0 ? 0 : yawBin;
 
                 // pitch
                 int32_t pitchBin = calorimeterPos.y() * static_cast<float_X>(numBinsPitch);
                 // catch out-of-range values
-                pitchBin = pitchBin >= numBinsPitch ? numBinsPitch - 1 : pitchBin;
+                pitchBin = pitchBin >= static_cast<int32_t>(numBinsPitch) ? numBinsPitch - 1 : pitchBin;
                 pitchBin = pitchBin < 0 ? 0 : pitchBin;
 
                 // energy
